(Cosmos/Tron): use same input currency for both delegation flow modals
